+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Update SCTK to 0.18
 - Fix active polling of the clipboard each 50ms
 - Fix freeze when copying data larger than the pipe buffer size
+- Accept text/plain mime type as a fallback
 
 ## 0.6.6 -- 2022-06-20